Abstract Blocks into new collections base
