+++ src/abstracts.ts
@@ -1,7 +1,3 @@
-export interface Queryable {
-    query(selector: string);
-}
-
 export interface AssetFetcher {
     parse(source: string | number): Promise<JSON>;
 }
